Move parsedVersion to separate file
